+++ clang/lib/ASTImporter/ASTImporterLookupTable.cpp
@@ -11,7 +11,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "clang/AST/ASTImporterLookupTable.h"
+#include "clang/ASTImporter/ASTImporterLookupTable.h"
 #include "clang/AST/Decl.h"
 #include "clang/AST/RecursiveASTVisitor.h"
 #include "llvm/Support/FormatVariadic.h"
