+++ clang/lib/ASTImporter/ExternalASTMerger.cpp
@@ -16,7 +16,7 @@
 #include "clang/AST/DeclCXX.h"
 #include "clang/AST/DeclObjC.h"
 #include "clang/AST/DeclTemplate.h"
-#include "clang/AST/ExternalASTMerger.h"
+#include "clang/ASTImporter/ExternalASTMerger.h"
 
 using namespace clang;
