+++ clang/unittests/AST/ASTImporterFixtures.cpp
@@ -13,8 +13,6 @@
 
 #include "ASTImporterFixtures.h"
 
-#include "clang/AST/ASTImporter.h"
-#include "clang/AST/ASTImporterSharedState.h"
 #include "clang/Frontend/ASTUnit.h"
 #include "clang/Tooling/Tooling.h"
